% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_fix.R
\name{desfix_bayes}
\alias{desfix_bayes}
\title{Bayesian log normal models}
\usage{
desfix_bayes(
  y,
  L_m,
  U_m,
  L_cv,
  U_cv,
  L_m_inc = NA,
  U_m_inc = NA,
  data_hist = NULL,
  n_min_borrow = 3,
  bayes_model = c("independent", "same_cv", "monotone"),
  ...
)
}
\description{
For y, L_m and U_m
}
