% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_fix.R
\name{desfix_beta_binomial_desity}
\alias{desfix_beta_binomial_desity}
\title{Plot binomial density}
\usage{
desfix_beta_binomial_desity(x, n, alpha, beta, cut, prior_post)
}
\description{
Plot binomial density
}
