% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_fix.R
\name{desfix_generate_patients}
\alias{desfix_generate_patients}
\title{Simulate patients}
\usage{
desfix_generate_patients(mean_log, sd_log, dlt_rate, n_patients)
}
\description{
the lengths of dose_level, mean_log, sd_log and dlt_rate must be the same
}
\details{
right now let n_patients to be the same across dose levels
}
