% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_fix.R
\name{desfix_kernel_density}
\alias{desfix_kernel_density}
\title{Plot kernel density}
\usage{
desfix_kernel_density(x, cut, xlab_name, para_notation, percent, from, to)
}
\description{
Plot kernel density
}
