% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_fix.R
\name{desfix_summary}
\alias{desfix_summary}
\title{Summarize results}
\usage{
desfix_summary(results, lst_design, ...)
}
\description{
Dose escalation algorithm
}
\details{
0:stay

-12:stop based on predicted and posterior mean FIX given 6 subjects
-13:stop based on predicted and posterior mean FIX given 3~5 subjects
-14:stop due to both over posterior DLT given >=2 subjects and over observed FIX for any subject
-15:stop due to over observed FIX for any subjects
-16:stop due to over posterior DLT given >=2 subjects
No stop rule for DLT when only 1 subject; Has stop rule for FIX for any subject.

11: escalate due to very low observed FIX of the 1st subjects
12: escalate based on predicted and posterior mean FIX given 6 subjects
13: escalate based on predicted and posterior mean FIX given 3~5 subjects

2:selected based on predicted and posterior mean FIX given 6 subjects
}
