% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_fix.R
\name{get_guarded}
\alias{get_guarded}
\title{Design point}
\usage{
get_guarded(
  dta_fix,
  dta_dlt,
  fix_interval_ind = c(5, 150),
  dlt_thresh_r = 0.15,
  dlt_thresh_c = 0.8,
  dlt_prior = c(0.1, 0.9)
)
}
\description{
Whether the dose is safe
}
