% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_package.R
\docType{package}
\name{rankpts-package}
\alias{rankpts-package}
\alias{rankpts}
\title{The 'rankpts' package.}
\description{
An R Package for Treatment Effect Evaluation Based On Ranking Patients
}
\author{
\strong{Maintainer}: Chenguang Wang \email{chenguang.wang.0517@gmail.com}

}
