% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpt_door.R
\name{rpt_rank_matrix}
\alias{rpt_rank_matrix}
\title{Prepare data for ranking patients}
\usage{
rpt_rank_matrix(dta, dta_rules)

rpt_rank_matrix(dta, dta_rules)
}
\description{
For all variable, a bigger value means a better outcome

For all variable, a bigger value means a better outcome
}
