% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpt_tls_stan_shiny.R
\name{rpt_shiny}
\alias{rpt_shiny}
\title{Run Web-Based application}
\usage{
rpt_shiny(appname = "rank", pkgname = "rankpts")
}
\description{
Call Shiny to run \code{statidea} as a web-based application.
}
\details{
A web browser will be brought up for users to access the GUI
}
