% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_phasei.R
\name{stb_tl_p1_simu_cohort}
\alias{stb_tl_p1_simu_cohort}
\title{Simulate a cohort of patients}
\usage{
stb_tl_p1_simu_cohort(
  cur_n,
  cur_prob,
  inx = 1,
  par_enroll,
  enroll_shift = 0,
  dlt_days = 28,
  ...
)
}
\description{
Simulate a cohort of patients
}
