% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_enroll.R
\name{stb_tl_simu_enroll_cohort}
\alias{stb_tl_simu_enroll_cohort}
\title{Simulate Enrollment Time}
\usage{
stb_tl_simu_enroll_cohort(
  n_pt = 3,
  par_enroll = list(type = "by_duration", pt_dur_mth = 24),
  date_bos = "2022-01-01",
  mth_to_days = 30.4,
  ...
)
}
\description{
Simulate Enrollment Time
}
